+++ pkgs/checks/lib/src/extensions/core.dart
@@ -35,8 +35,10 @@ extension CoreChecks<T> on Subject<T> {
   /// }
   /// ```
   @meta.useResult
-  Subject<R> has<R>(R Function(T) extract, String name) {
-    return context.nest(() => ['has $name'], (value) {
+  Subject<R> has<R>(R Function(T) extract, String name) => context.nest(
+    () => ['has $name'],
+    addPredicate: (predicateNoun) => 'has $name: $predicateNoun',
+    (value) {
       try {
         return Extracted.value(extract(value));
       } catch (e, st) {
@@ -50,8 +52,8 @@ extension CoreChecks<T> on Subject<T> {
           ],
         );
       }
-    });
-  }
+    },
+  );
 
   /// Applies the expectations invoked in [condition] to this subject.
   ///
@@ -62,11 +64,11 @@ extension CoreChecks<T> on Subject<T> {
   /// ```
   /// check(something)
   ///   ..has((s) => s.foo, 'foo').equals(expectedFoo)
-  ///   ..has((s) => s.bar, 'bar').which((b) => b
+  ///   ..has((s) => s.bar, 'bar').which(.it()
   ///     ..isLessThan(10)
   ///     ..isGreaterThan(0));
   /// ```
-  void which(Condition<T> condition) => condition(this);
+  void which(Condition<T> condition) => condition.applySync(this);
 
   /// Check that the expectations invoked in [condition] are not satisfied by
   /// this value.
@@ -74,11 +76,11 @@ extension CoreChecks<T> on Subject<T> {
   /// Asynchronous expectations are not allowed in [condition].
   void not(Condition<T> condition) {
     context.expect(
-      () => ['is not a value that:', ...indent(describe(condition))],
+      () => ['is not a value that:', ...indent(condition.describeSync())],
       (actual) {
-        if (softCheck(actual, condition) != null) return null;
+        if (condition.softCheckSync(actual) != null) return null;
         return Rejection(
-          which: ['is a value that: ', ...indent(describe(condition))],
+          which: ['is a value that: ', ...indent(condition.describeSync())],
         );
       },
     );
@@ -93,7 +95,7 @@ extension CoreChecks<T> on Subject<T> {
       () => prefixFirst('matches any condition in ', literal(conditions)),
       (actual) {
         for (final condition in conditions) {
-          if (softCheck(actual, condition) == null) return null;
+          if (condition.softCheckSync(actual) == null) return null;
         }
         return Rejection(which: ['did not match any condition']);
       },
@@ -103,13 +105,13 @@ extension CoreChecks<T> on Subject<T> {
   /// Expects that the value is assignable to type [T].
   ///
   /// If the value is a [T], returns a [Subject] for further expectations.
-  Subject<R> isA<R>() {
-    return context.nest<R>(() => ['is a $R'], (actual) {
+  Subject<R> isA<R>([Condition<R>? and]) {
+    return context.nest<R>(() => ['is a $R'], atSameLevel: true, (actual) {
       if (actual is! R) {
         return Extracted.rejection(which: ['Is a ${actual.runtimeType}']);
       }
       return Extracted.value(actual);
-    }, atSameLevel: true);
+    }, nestedCondition: and);
   }
 
   /// Expects that the value is not assignable to type [R].
@@ -124,20 +126,26 @@ extension CoreChecks<T> on Subject<T> {
 
   /// Expects that the value is equal to [other] according to [operator ==].
   void equals(T other) {
-    context.expect(() => prefixFirst('equals ', literal(other)), (actual) {
-      if (actual == other) return null;
-      return Rejection(which: ['are not equal']);
-    });
+    context.expect(
+      () => prefixFirst('equals ', literal(other)),
+      predicateNoun: () => literal(other).singleOrNull,
+      (actual) {
+        if (actual == other) return null;
+        return Rejection(which: ['is not equal']);
+      },
+    );
   }
 
   /// Expects that the value is [identical] to [other].
   void identicalTo(T other) {
-    context.expect(() => prefixFirst('is identical to ', literal(other)), (
-      actual,
-    ) {
-      if (identical(actual, other)) return null;
-      return Rejection(which: ['is not identical']);
-    });
+    context.expect(
+      () => prefixFirst('is identical to ', literal(other)),
+      predicateNoun: () => literal(other).singleOrNull,
+      (actual) {
+        if (identical(actual, other)) return null;
+        return Rejection(which: ['is not identical']);
+      },
+    );
   }
 }
 
@@ -145,6 +153,7 @@ extension BoolChecks on Subject<bool> {
   void isTrue() {
     context.expect(
       () => ['is true'],
+      predicateNoun: () => 'true',
       (actual) => actual
           ? null // force coverage
           : Rejection(),
@@ -154,6 +163,7 @@ extension BoolChecks on Subject<bool> {
   void isFalse() {
     context.expect(
       () => ['is false'],
+      predicateNoun: () => 'false',
       (actual) => !actual
           ? null // force coverage
           : Rejection(),
@@ -162,17 +172,19 @@ extension BoolChecks on Subject<bool> {
 }
 
 extension NullableChecks<T> on Subject<T?> {
-  Subject<T> isNotNull() {
-    return context.nest<T>(() => ['is not null'], (actual) {
+  Subject<T> isNotNull([Condition<T>? and]) {
+    return context.nest<T>(() => ['is not null'], atSameLevel: true, (actual) {
       if (actual == null) return Extracted.rejection();
       return Extracted.value(actual);
-    }, atSameLevel: true);
+    }, nestedCondition: and);
   }
 
   void isNull() {
-    context.expect(() => const ['is null'], (actual) {
-      if (actual != null) return Rejection();
-      return null;
+    context.expect(() => const ['is null'], predicateNoun: () => 'null', (
+      actual,
+    ) {
+      if (actual == null) return null;
+      return Rejection();
     });
   }
 }
@@ -180,20 +192,29 @@ extension NullableChecks<T> on Subject<T?> {
 extension ComparableChecks<T> on Subject<Comparable<T>> {
   /// Expects that this value is greater than [other].
   void isGreaterThan(T other) {
-    context.expect(() => prefixFirst('is greater than ', literal(other)), (
-      actual,
-    ) {
-      if (actual.compareTo(other) > 0) return null;
-      return Rejection(
-        which: prefixFirst('is not greater than ', literal(other)),
-      );
-    });
+    context.expect(
+      () => prefixFirst('is greater than ', literal(other)),
+      predicateNoun: () {
+        final l = literal(other).singleOrNull;
+        return l != null ? 'a value > $l' : null;
+      },
+      (actual) {
+        if (actual.compareTo(other) > 0) return null;
+        return Rejection(
+          which: prefixFirst('is not greater than ', literal(other)),
+        );
+      },
+    );
   }
 
   /// Expects that this value is greater than or equal to [other].
   void isGreaterOrEqual(T other) {
     context.expect(
       () => prefixFirst('is greater than or equal to ', literal(other)),
+      predicateNoun: () {
+        final l = literal(other).singleOrNull;
+        return l != null ? 'a value >= $l' : null;
+      },
       (actual) {
         if (actual.compareTo(other) >= 0) return null;
         return Rejection(
@@ -208,18 +229,29 @@ extension ComparableChecks<T> on Subject<Comparable<T>> {
 
   /// Expects that this value is less than [other].
   void isLessThan(T other) {
-    context.expect(() => prefixFirst('is less than ', literal(other)), (
-      actual,
-    ) {
-      if (actual.compareTo(other) < 0) return null;
-      return Rejection(which: prefixFirst('is not less than ', literal(other)));
-    });
+    context.expect(
+      () => prefixFirst('is less than ', literal(other)),
+      predicateNoun: () {
+        final l = literal(other).singleOrNull;
+        return l != null ? 'a value < $l' : null;
+      },
+      (actual) {
+        if (actual.compareTo(other) < 0) return null;
+        return Rejection(
+          which: prefixFirst('is not less than ', literal(other)),
+        );
+      },
+    );
   }
 
   /// Expects that this value is less than or equal to [other].
   void isLessOrEqual(T other) {
     context.expect(
       () => prefixFirst('is less than or equal to ', literal(other)),
+      predicateNoun: () {
+        final l = literal(other).singleOrNull;
+        return l != null ? 'a value <= $l' : null;
+      },
       (actual) {
         if (actual.compareTo(other) <= 0) return null;
         return Rejection(