+++ pkgs/checks/test/describe_test.dart
@@ -3,21 +3,24 @@
 // BSD-style license that can be found in the LICENSE file.
 
 import 'package:checks/checks.dart';
-import 'package:checks/context.dart';
 import 'package:test/scaffolding.dart';
+
+import 'test_shared.dart';
 
 void main() {
   group('describe', () {
     test('succeeds for empty conditions', () {
-      check(describe((_) {})).isEmpty();
+      check(Condition.it<void>()).hasSyncDescription().isEmpty();
     });
     test('includes condition clauses', () {
-      check(describe((it) => it.equals(1))).deepEquals(['  equals <1>']);
+      check(
+        Condition.it<int>()..equals(1),
+      ).hasSyncDescription().deepEquals(['  equals <1>']);
     });
     test('includes nested clauses', () {
-      check(
-        describe<String>((it) => it.length.equals(1)),
-      ).deepEquals(['  has length that:', '    equals <1>']);
+      check(Condition.it<String>()..length.equals(1))
+          .hasSyncDescription()
+          .deepEquals(['  has length that:', '    equals <1>']);
     });
   });
 }