+++ pkgs/checks/test/extensions/core_test.dart
@@ -9,60 +9,74 @@ import '../test_shared.dart';
 
 void main() {
   group('TypeChecks', () {
-    test('isA', () {
-      check(1).isA<int>();
-
-      check(1).isRejectedBy((it) => it.isA<String>(), which: ['Is a int']);
+    group('isA', () {
+      test('happy case', () {
+        check(1).isA<int>();
+      });
+      test('failure case', () {
+        check(1).isRejectedBy(.it()..isA<String>(), which: ['Is a int']);
+      });
+      test('evaluates condition', () {
+        check(1).isRejectedBy(
+          .it()..isA<int>(.it()..isGreaterThan(2)),
+          which: ['is not greater than <2>'],
+        );
+      });
+      test('returns valid subject', () {
+        check(1).isA<int>().isGreaterThan(0);
+      });
     });
     test('isNotA', () {
       check(1).isNotA<String>();
 
-      check(1).isRejectedBy((it) => it.isNotA<int>(), which: ['is a int']);
-      check(1).isRejectedBy((it) => it.isNotA<num>(), which: ['is a num']);
+      check(1).isRejectedBy(.it()..isNotA<int>(), which: ['is a int']);
+      check(1).isRejectedBy(.it()..isNotA<num>(), which: ['is a num']);
     });
   });
   group('HasField', () {
-    test('has', () {
-      check(1).has((v) => v.isOdd, 'isOdd').isTrue();
-
-      check(null).isRejectedBy(
-        (it) => it.has((v) {
-          Error.throwWithStackTrace(
-            UnimplementedError(),
-            StackTrace.fromString('fake trace'),
-          );
-        }, 'foo').isNotNull(),
-        which: [
-          'threw while trying to read foo: <UnimplementedError> at:',
-          '  fake trace',
-        ],
-      );
+    group('has', () {
+      test('happy case', () {
+        check(1).has((v) => v.isOdd, 'isOdd').isTrue();
+      });
+      test('failure case', () {
+        check(null).isRejectedBy(
+          .it()
+            ..has((v) {
+              Error.throwWithStackTrace(
+                UnimplementedError(),
+                StackTrace.fromString('fake trace'),
+              );
+            }, 'foo').isNotNull(),
+          which: [
+            'threw while trying to read foo: <UnimplementedError> at:',
+            '  fake trace',
+          ],
+        );
+      });
+      test('returns valid subject', () {
+        check(1).has((v) => v.isOdd, 'isOdd').isTrue();
+      });
     });
 
     test('which', () {
-      check(true).which((it) => it.isTrue());
+      check(true).which(.it()..isTrue());
     });
 
     test('not', () {
-      check(false).not((it) => it.isTrue());
+      check(false).not(.it()..isTrue());
       check(true).isRejectedBy(
-        (it) => it.not((it) => it.isTrue()),
+        .it()..not(.it()..isTrue()),
         which: ['is a value that: ', '    is true'],
       );
     });
 
     group('anyOf', () {
       test('succeeds for happy case', () {
-        check(
-          -10,
-        ).anyOf([(it) => it.isGreaterThan(1), (it) => it.isLessThan(-1)]);
+        check(-10).anyOf([.it()..isGreaterThan(1), .it()..isLessThan(-1)]);
       });
       test('rejects values that do not satisfy any condition', () {
         check(0).isRejectedBy(
-          (it) => it.anyOf([
-            (it) => it.isGreaterThan(1),
-            (it) => it.isLessThan(-1),
-          ]),
+          .it()..anyOf([.it()..isGreaterThan(1), .it()..isLessThan(-1)]),
           which: ['did not match any condition'],
         );
       });
@@ -73,13 +87,13 @@ void main() {
     test('isTrue', () {
       check(true).isTrue();
 
-      check(false).isRejectedBy((it) => it.isTrue());
+      check(false).isRejectedBy(.it()..isTrue());
     });
 
     test('isFalse', () {
       check(false).isFalse();
 
-      check(true).isRejectedBy((it) => it.isFalse());
+      check(true).isRejectedBy(.it()..isFalse());
     });
   });
 
@@ -87,26 +101,36 @@ void main() {
     test('equals', () {
       check(1).equals(1);
 
-      check(1).isRejectedBy((it) => it.equals(2), which: ['are not equal']);
+      check(1).isRejectedBy(.it()..equals(2), which: ['is not equal']);
     });
     test('identical', () {
       check(1).identicalTo(1);
 
-      check(
-        1,
-      ).isRejectedBy((it) => it.identicalTo(2), which: ['is not identical']);
+      check(1).isRejectedBy(.it()..identicalTo(2), which: ['is not identical']);
     });
   });
   group('NullabilityChecks', () {
-    test('isNotNull', () {
-      check(1).isNotNull();
-
-      check(null).isRejectedBy((it) => it.isNotNull());
+    group('isNotNull', () {
+      test('happy case', () {
+        check(1).isNotNull();
+      });
+      test('failure case', () {
+        check(null).isRejectedBy(.it()..isNotNull());
+      });
+      test('evaluates condition', () {
+        check(1).isRejectedBy(
+          .it()..isNotNull(.it()..identicalTo(2)),
+          which: ['is not identical'],
+        );
+      });
+      test('returns valid subject', () {
+        check(1).isNotNull().identicalTo(1);
+      });
     });
     test('isNull', () {
       check(null).isNull();
 
-      check(1).isRejectedBy((it) => it.isNull());
+      check(1).isRejectedBy(.it()..isNull());
     });
   });
 
@@ -119,13 +143,13 @@ void main() {
       });
       test('fails for equal', () {
         check(const Duration(seconds: 10)).isRejectedBy(
-          (it) => it.isGreaterThan(const Duration(seconds: 10)),
+          .it()..isGreaterThan(const Duration(seconds: 10)),
           which: ['is not greater than <0:00:10.000000>'],
         );
       });
       test('fails for less', () {
         check(const Duration(seconds: 10)).isRejectedBy(
-          (it) => it.isGreaterThan(const Duration(seconds: 50)),
+          .it()..isGreaterThan(const Duration(seconds: 50)),
           which: ['is not greater than <0:00:50.000000>'],
         );
       });
@@ -143,7 +167,7 @@ void main() {
       });
       test('fails for less', () {
         check(const Duration(seconds: 10)).isRejectedBy(
-          (it) => it.isGreaterOrEqual(const Duration(seconds: 50)),
+          .it()..isGreaterOrEqual(const Duration(seconds: 50)),
           which: ['is not greater than or equal to <0:00:50.000000>'],
         );
       });
@@ -156,13 +180,13 @@ void main() {
       });
       test('fails for equal', () {
         check(const Duration(seconds: 10)).isRejectedBy(
-          (it) => it.isLessThan(const Duration(seconds: 10)),
+          .it()..isLessThan(const Duration(seconds: 10)),
           which: ['is not less than <0:00:10.000000>'],
         );
       });
       test('fails for greater', () {
         check(const Duration(seconds: 50)).isRejectedBy(
-          (it) => it.isLessThan(const Duration(seconds: 10)),
+          .it()..isLessThan(const Duration(seconds: 10)),
           which: ['is not less than <0:00:10.000000>'],
         );
       });
@@ -180,7 +204,7 @@ void main() {
       });
       test('fails for greater', () {
         check(const Duration(seconds: 10)).isRejectedBy(
-          (it) => it.isLessOrEqual(const Duration(seconds: 1)),
+          .it()..isLessOrEqual(const Duration(seconds: 1)),
           which: ['is not less than or equal to <0:00:01.000000>'],
         );
       });